Added download-all feature without instanceconfig
The button to download all the files in a thread only appears when the
page is a thread, such as treeview.
I added the download-all button next
to the treeview one.
I modified stylesheets/v8ch.css to include the id
#download-all.
I DIDN'T modified inc/instance-config.php to include the
download-all.js and jszip.min.js. That goes in the next commit.

I added the files js/jszip.min.js
which is a library to work with zip files in JS and I added
js/download-all.js which goes through all files, adds them to a zip and
then gives it to the user.