Change from fopen/flock to dio_open/dio_fcntl in file_write
8chan uses NFS and flock() does not work over NFS. See http://0pointer.de/blog/projects/locking.html for more information.

Without proper file locking, race conditions are possible in ?/settings and other pages. The one in ?/settings is particularly bad, too many successive writes can cause a PHP file with bad syntax to be written which breaks an entire board and many scripts that call openBoard().

You need to install the dio.so module if you merge this commit.