Fix "double dorito" bug (post filter runs twice)
Note: This broke post-hover when expand-too-long is used. It's necessary to either come up with an event other than new_post for this (perhaps change_post), or have post-hover.js and other safe scripts expose a method that can be called here. I think the new event is a better idea though.