Added advanced version of Download All.
This was tested using a userscript. I can't run infinity, so that's the only way I can test it. Should work, however.

This is an advanced version of the Download All feature (I saw there was already a download-all.js when I was about to commit this, so I guess this is the advanced version). It has a menu that shows every image, and every image has a checkbox that decides if that should be downloaded or not. There are buttons that check all, uncheck all, and reverse the check. You type in the name of the zip (default is the name of the thread and .zip), press Download, and it creates the zip file and downloads it.